+++ src/main/java/subway/section/SectionDao.java
@@ -1,6 +1,5 @@
 package subway.section;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
