+++ src/main/java/subway/station/StationController.java
@@ -1,11 +1,8 @@
 package subway.station;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.exceptions.DuplicateStationNameException;
-import subway.exceptions.InvalidStationArgumentException;
 
 import java.net.URI;
 import java.util.ArrayList;
@@ -14,12 +11,11 @@ import java.util.List;
 @RestController
 @RequestMapping("stations")
 public class StationController {
-    @Autowired
+
     private StationService stationService;
 
-    @ExceptionHandler({DuplicateStationNameException.class, InvalidStationArgumentException.class})
-    public ResponseEntity<String> errorHandler(Exception e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
+    public StationController(StationService stationService) {
+        this.stationService = stationService;
     }
 
     @PostMapping
