+++ src/main/java/subway/station/StationService.java
@@ -1,8 +1,6 @@
 package subway.station;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import subway.exceptions.InvalidStationArgumentException;
 
 import java.util.List;
